Corrected issues in machine_state

+++ src/mips-core/machine_state.cpp
@@ -52,7 +52,6 @@ void MachineState::set_register(uint8_t index, uint32_t value) {
 }
 
 uint32_t MachineState::read_word(uint32_t address) const {
-    // bound check
     if (address > memory.size() - 4) {
         throw std::out_of_range("Memory read out of bounds");
     }
@@ -62,8 +61,7 @@ uint32_t MachineState::read_word(uint32_t address) const {
     return value;
 }
 
-uint32_t MachineState::read_half_word_signed(uint32_t address) const {
-    // bound check
+int16_t MachineState::read_half_word_signed(uint32_t address) const {
     if (address > memory.size() - 2) {
         throw std::out_of_range("Memory read out of bounds");
     }
@@ -73,8 +71,7 @@ uint32_t MachineState::read_half_word_signed(uint32_t address) const {
     return value;
 }
 
-uint32_t MachineState::read_half_word_unsigned(uint32_t address) const {
-    // bound check
+uint16_t MachineState::read_half_word_unsigned(uint32_t address) const {
     if (address > memory.size() - 2) {
         throw std::out_of_range("Memory read out of bounds");
     }
@@ -84,8 +81,7 @@ uint32_t MachineState::read_half_word_unsigned(uint32_t address) const {
     return value;
 }
 
-uint32_t MachineState::read_byte_signed(uint32_t address) const {
-    // bound check
+int8_t MachineState::read_byte_signed(uint32_t address) const {
     if (address >= memory.size()) {
         throw std::out_of_range("Memory read out of bounds");
     }
@@ -95,8 +91,7 @@ uint32_t MachineState::read_byte_signed(uint32_t address) const {
     return value;
 }
 
-uint32_t MachineState::read_byte_unsigned(uint32_t address) const {
-    // bound check
+uint8_t MachineState::read_byte_unsigned(uint32_t address) const {
     if (address >= memory.size()) {
         throw std::out_of_range("Memory read out of bounds");
     }
@@ -107,7 +102,6 @@ uint32_t MachineState::read_byte_unsigned(uint32_t address) const {
 }
 
 void MachineState::write_word(uint32_t address, uint32_t value) {
-    // bound check
     if (address > memory.size() - 4) {
         throw std::out_of_range("Memory write out of bounds");
     }
@@ -116,7 +110,6 @@ void MachineState::write_word(uint32_t address, uint32_t value) {
 }
 
 void MachineState::write_half_word(uint32_t address, uint16_t value) {
-    // bound check
     if (address > memory.size() - 2) {
         throw std::out_of_range("Memory write out of bounds");
     }
@@ -125,7 +118,6 @@ void MachineState::write_half_word(uint32_t address, uint16_t value) {
 }
 
 void MachineState::write_byte(uint32_t address, uint8_t value) {
-    // bound check
     if (address >= memory.size()) {
         throw std::out_of_range("Memory write out of bounds");
     }
